+++ src/loader.js
@@ -10,7 +10,9 @@ import SingleEntryPlugin from 'webpack/lib/SingleEntryPlugin';
 import LimitChunkCountPlugin from 'webpack/lib/optimize/LimitChunkCountPlugin';
 import validateOptions from 'schema-utils';
 
-import schema from './options.json';
+import CssDependency from './CssDependency';
+
+import schema from './loader-options.json';
 
 const MODULE_TYPE = 'css/mini-extract';
 const pluginName = 'mini-css-extract-plugin';
@@ -36,7 +38,7 @@ function hotLoader(content, context) {
   `;
 }
 
-const exec = (loaderContext, code, filename) => {
+function evalModuleCode(loaderContext, code, filename) {
   const module = new NativeModule(filename, loaderContext);
 
   module.paths = NativeModule._nodeModulePaths(loaderContext.context); // eslint-disable-line no-underscore-dangle
@@ -44,9 +46,9 @@ const exec = (loaderContext, code, filename) => {
   module._compile(code, filename); // eslint-disable-line no-underscore-dangle
 
   return module.exports;
-};
+}
 
-const findModuleById = (modules, id) => {
+function findModuleById(modules, id) {
   for (const module of modules) {
     if (module.id === id) {
       return module;
@@ -54,7 +56,7 @@ const findModuleById = (modules, id) => {
   }
 
   return null;
-};
+}
 
 export function pitch(request) {
   const options = loaderUtils.getOptions(this) || {};
@@ -69,7 +71,7 @@ export function pitch(request) {
     return;
   }
 
-  const childFilename = '*'; // eslint-disable-line no-path-concat
+  const childFilename = '*';
   const publicPath =
     typeof options.publicPath === 'string'
       ? options.publicPath === '' || options.publicPath.endsWith('/')
@@ -95,8 +97,6 @@ export function pitch(request) {
   );
   new LimitChunkCountPlugin({ maxChunks: 1 }).apply(childCompiler);
 
-  // We set loaderContext[MODULE_TYPE] = false to indicate we already in
-  // a child compiler so we don't spawn another child compilers from there.
   childCompiler.hooks.thisCompilation.tap(
     `${pluginName} loader`,
     (compilation) => {
@@ -105,7 +105,6 @@ export function pitch(request) {
         (loaderContext, module) => {
           // eslint-disable-next-line no-param-reassign
           loaderContext.emitFile = this.emitFile;
-          loaderContext[MODULE_TYPE] = false; // eslint-disable-line no-param-reassign
 
           if (module.request === request) {
             // eslint-disable-next-line no-param-reassign
@@ -140,6 +139,27 @@ export function pitch(request) {
   const callback = this.async();
 
   childCompiler.runAsChild((err, entries, compilation) => {
+    const addDependencies = (dependencies) => {
+      if (!Array.isArray(dependencies) && dependencies != null) {
+        throw new Error(
+          `Exported value was not extracted as an array: ${JSON.stringify(
+            dependencies
+          )}`
+        );
+      }
+
+      const identifierCountMap = new Map();
+
+      for (const dependency of dependencies) {
+        const count = identifierCountMap.get(dependency.identifier) || 0;
+
+        this._module.addDependency(
+          new CssDependency(dependency, dependency.context, count)
+        );
+        identifierCountMap.set(dependency.identifier, count + 1);
+      }
+    };
+
     if (err) {
       return callback(err);
     }
@@ -160,35 +180,43 @@ export function pitch(request) {
       return callback(new Error("Didn't get a result from child compiler"));
     }
 
-    let text;
     let locals;
 
     try {
-      text = exec(this, source, request);
-      locals = text && text.locals;
-      if (!Array.isArray(text)) {
-        text = [[null, text]];
+      let dependencies;
+      let exports = evalModuleCode(this, source, request);
+      // eslint-disable-next-line no-underscore-dangle
+      exports = exports.__esModule ? exports.default : exports;
+      locals = exports && exports.locals;
+      if (!Array.isArray(exports)) {
+        dependencies = [[null, exports]];
       } else {
-        text = text.map((line) => {
-          const module = findModuleById(compilation.modules, line[0]);
+        dependencies = exports.map(([id, content, media, sourceMap]) => {
+          const module = findModuleById(compilation.modules, id);
 
           return {
             identifier: module.identifier(),
-            content: line[1],
-            media: line[2],
-            sourceMap: line[3],
+            context: module.context,
+            content,
+            media,
+            sourceMap,
           };
         });
       }
-      this[MODULE_TYPE](text);
+      addDependencies(dependencies);
     } catch (e) {
       return callback(e);
     }
 
-    let resultSource = `// extracted by ${pluginName}`;
+    const esModule =
+      typeof options.esModule !== 'undefined' ? options.esModule : false;
     const result = locals
-      ? `\nmodule.exports = ${JSON.stringify(locals)};`
+      ? `\n${esModule ? 'export default' : 'module.exports ='} ${JSON.stringify(
+          locals
+        )};`
       : '';
+
+    let resultSource = `// extracted by ${pluginName}`;
 
     resultSource += options.hmr
       ? hotLoader(result, { context: this.context, options, locals })