+++ test/cases/composes-async/index.js
@@ -1,2 +1,4 @@
-import('./async-1.css');
-import('./async-2.css');
+/* eslint-disable-next-line no-unused-expressions */
+import(/* webpackChunkName: "async-1" */ './async-1.css');
+/* eslint-disable-next-line no-unused-expressions */
+import(/* webpackChunkName: "async-2" */ './async-2.css');