+++ test/cases/hmr/expected/webpack-5/main.js
@@ -1,0 +1,7 @@
+/******/ (() => { // webpackBootstrap
+/******/ 	/************************************************************************/
+// extracted by mini-css-extract-plugin
+    if(false) { var cssReload; }
+  
+/******/ })()
+;