+++ test/cases/split-chunks-single/entry1.js
@@ -3,5 +3,7 @@ import './e1.css';
 import './e2.css';
 import './f.css';
 
+/* eslint-disable-next-line no-unused-expressions */
 import('./chunk1');
+/* eslint-disable-next-line no-unused-expressions */
 import('./chunk2');