+++ test/manual/webpack.config.js
@@ -1,7 +1,18 @@
 const Self = require('../../');
 
+const ENABLE_HMR =
+  typeof process.env.ENABLE_HMR !== 'undefined'
+    ? Boolean(process.env.ENABLE_HMR)
+    : false;
+
+const ENABLE_ES_MODULE =
+  typeof process.env.ES_MODULE !== 'undefined'
+    ? Boolean(process.env.ES_MODULE)
+    : false;
+
 module.exports = {
   mode: 'development',
+  devtool: 'cheap-source-map',
   output: {
     chunkFilename: '[contenthash].js',
     publicPath: '/dist/',
@@ -11,7 +22,42 @@ module.exports = {
     rules: [
       {
         test: /\.css$/,
-        use: ['style-loader', Self.loader, 'css-loader'],
+        exclude: [/\.module\.css$/i],
+        use: [
+          {
+            loader : 'style-loader',
+          },
+          {
+            loader: Self.loader,
+            options: {
+              hmr: ENABLE_HMR,
+            },
+          },
+          {
+            loader: 'css-loader',
+            options: {
+              esModule: ENABLE_ES_MODULE,
+            },
+          },
+        ],
+      },
+      {
+        test: /\.module\.css$/i,
+        use: [
+          {
+            loader: Self.loader,
+            options: {
+              hmr: ENABLE_HMR,
+            },
+          },
+          {
+            loader: 'css-loader',
+            options: {
+              modules: true,
+              esModule: ENABLE_ES_MODULE,
+            },
+          },
+        ],
       },
     ],
   },
@@ -19,7 +65,7 @@ module.exports = {
     new Self({
       filename: '[name].css',
       chunkFilename: '[contenthash].css',
-      disableAsync({ module }) {
+      disableExtract({ module }) {
         let ret = false;
         if (module.content.indexOf('async-disabled') > -1) {
           ret = true;